refactor: ROUTES 상수 분리
