+++ src/components/auth/ProtectedRoute.tsx
@@ -1,6 +1,6 @@
 import type { ReactElement } from "react";
 import { Navigate, useLocation } from "react-router-dom";
-import { ROUTES } from "@/constants/endpoint";
+import { ROUTES } from "@/constants/routes";
 import { tokenStore } from "@/utils/tokenStore";
 
 interface ProtectedRouteProps {
