+++ src/constants/endpoint.ts
@@ -22,15 +22,4 @@ const ENDPOINT = {
 
 export const WS_BASE_URL = API_BASE_URL.replace(/^http/i, "ws");
 
-export const ROUTES = {
-  APP_ROOT: "/",
-  SIGNUP: "/signup",
-  LOGIN: "/login",
-  DECISIONS: "/decisions",
-  MEETING: "/meeting",
-  MEETING_DETAIL: "/meeting/:meetingId",
-  GIT: "/git",
-  SETTINGS: "/settings",
-} as const;
-
 export default ENDPOINT;
