+++ src/constants/routes.ts
@@ -1,0 +1,10 @@
+export const ROUTES = {
+  APP_ROOT: "/",
+  SIGNUP: "/signup",
+  LOGIN: "/login",
+  DECISIONS: "/decisions",
+  MEETING: "/meeting",
+  MEETING_DETAIL: "/meeting/:meetingId",
+  GIT: "/git",
+  SETTINGS: "/settings",
+} as const;
